Change functionality of delete menuItems
If a menuItem is being deleted from the MasterMenu on MenuItemsPage it will delete the menuItem entirely.

If a menuItem is being deleted from another menu, it will only delete the menuItem from that menu.

Fixed an alert in MenuItemSwap that wouldn't've triggered and uncommented some code I was testing the functionality of.